refactored further with example
