+++ praetorian_cli/sdk/chariot.py
@@ -12,11 +12,13 @@ from praetorian_cli.sdk.entities.integrations import Integrations
 from praetorian_cli.sdk.entities.jobs import Jobs
 from praetorian_cli.sdk.entities.preseeds import Preseeds
 from praetorian_cli.sdk.entities.risks import Risks
-from praetorian_cli.sdk.entities.search import GLOBAL_FLAG, Search, Query, Node, Filter
+from praetorian_cli.sdk.entities.search import Search
 from praetorian_cli.sdk.entities.seeds import Seeds
 from praetorian_cli.sdk.entities.statistics import Statistics
 from praetorian_cli.sdk.entities.webhook import Webhook
 from praetorian_cli.sdk.keychain import Keychain
+from praetorian_cli.sdk.model.globals import GLOBAL_FLAG
+from praetorian_cli.sdk.model.query import QueryBuilderDirector, Query, isGraphType
 
 QUERY_LIMIT_SWITCH_POINT = 10
 LARGE_QUERY_LIMIT = 5000
@@ -44,9 +46,11 @@ class Chariot:
         final_resp = dict()
         
         # For large, graph db queries, we can override default small query limit using a query object
-        query = swap_query_type(params, pages)
-        if query:
-            return self.my_by_query(query, pages, params.get('global', False))
+        key = params.get('key', None)
+        if key and isGraphType(key):
+            query = QueryBuilderDirector().from_params(params)
+            return self.my_by_query(query, pages)
+        
         for _ in range(pages):
             resp = requests.get(self.url('/my'), params=params, headers=self.keychain.headers())
             process_failure(resp)
@@ -62,10 +66,10 @@ class Chariot:
 
         return final_resp
 
-    def my_by_query(self, query: Query, pages=1, global_=False) -> {}:
+    def my_by_query(self, query: Query, pages=1) -> {}:
         final_resp = dict()
         for _ in range(pages):
-            resp = requests.post(self.url('/my'), json=query.to_dict(), params={"global": global_}, headers=self.keychain.headers())
+            resp = requests.post(self.url('/my'), json=query.to_dict(), params=query.get_params(), headers=self.keychain.headers())
             process_failure(resp)
             resp = resp.json()
             extend(final_resp, resp)
@@ -76,7 +80,6 @@ class Chariot:
 
         if 'offset' in resp:
             final_resp['offset'] = resp['offset']
-
         return final_resp
 
     def post(self, type: str, body: dict, params: dict = {}):
@@ -167,21 +170,6 @@ class Chariot:
 
     def url(self, path: str):
         return self.keychain.base_url() + path
-
-
-def swap_query_type(params: dict, pages: int):
-    if pages >= QUERY_LIMIT_SWITCH_POINT:
-        key = params.get('key', None)
-        if key:
-            prefix_list = []
-            prefix_list.extend([prefix.value.lower() for prefix in Filter.Field])
-            prefix_list.extend([label.value.lower() for label in Node.Label])
-            key = key.removeprefix('#')
-            if any([key.startswith(prefix) for prefix in prefix_list]):
-                params['limit'] = LARGE_QUERY_LIMIT
-                return Query(params=params)
-    return None
-
 
 def process_failure(response):
     if not response.ok:
