+++ praetorian_cli/sdk/entities/risks.py
@@ -1,5 +1,5 @@
 from praetorian_cli.sdk.model.globals import Kind
-from praetorian_cli.sdk.entities.search import Relationship, Node, Query, risk_of_key, ASSET_NODE, ATTRIBUTE_NODE
+from praetorian_cli.sdk.model.query import Relationship, Node, Query, risk_of_key, ASSET_NODE, ATTRIBUTE_NODE, SimpleQueryBuilder
 
 
 class Risks:
@@ -99,11 +99,15 @@ class Risks:
 
     def affected_assets(self, key):
         # assets directly linked to the risk
-        to_this = Relationship(Relationship.Label.HAS_VULNERABILITY, target=risk_of_key(key))
-        query = Query(Node(ASSET_NODE, relationships=[to_this]))
+        query = (SimpleQueryBuilder()
+        .add_node_label(Node.Label.ASSET)
+        .add_relationship(Relationship.Label.HAS_VULNERABILITY, target=risk_of_key(key))
+        .build()
+        )
         assets, _ = self.api.search.by_query(query)
 
         # assets indirectly linked to the risk via an attribute
+        to_this = Relationship(Relationship.Label.HAS_VULNERABILITY, target=risk_of_key(key))
         attributes = Node(ATTRIBUTE_NODE, relationships=[to_this])
         to_attributes = Relationship(Relationship.Label.HAS_ATTRIBUTE, target=attributes)
         query = Query(Node(ASSET_NODE, relationships=[to_attributes]))
