+++ praetorian_cli/sdk/entities/search.py
@@ -1,141 +1,5 @@
-from enum import Enum
-
-from praetorian_cli.sdk.model.globals import Kind
-from praetorian_cli.sdk.model.utils import get_type_from_key
-EXACT_FLAG = {'exact': 'true'}
-DESCENDING_FLAG = {'desc': 'true'}
-GLOBAL_FLAG = {'global': 'true'}
-
-
-class Filter:
-    class Operator(Enum):
-        EQUAL = '='
-        CONTAINS = 'CONTAINS'
-        LESS_THAN = '<'
-        LARGER_THAN = '>'
-        STARTS_WITH = 'STARTS WITH'
-        ENDS_WITH = 'ENDS WITH'
-
-    class Field(Enum): # Chariot.API.My depends upon these fields
-        KEY = 'key'
-        DNS = 'dns'
-        NAME = 'name'
-        STATUS = 'status'
-        SOURCE = 'source'
-        CREATED = 'created'
-
-    def __init__(self, field: Field = None, operator: Operator = None, value: str = None, params: dict = None):
-        if params:
-            key = params.get('key', None)
-            if key:
-                if ':' in key:
-                    field = Filter.Field(key.split(':')[0])
-                    value = key.split(':')[1]
-                else:
-                    field = Filter.Field.KEY
-                    value = key
-                operator = Filter.Operator.STARTS_WITH
-        self.field = field
-        self.operator = operator
-        self.value = value
-
-    def to_dict(self) -> dict:
-        if self.field == None or self.operator == None or self.value == None:
-            return None
-        return dict(field=self.field.value, operator=self.operator.value, value=self.value)
-
-class Relationship:
-    class Label(Enum):
-        HAS_VULNERABILITY = 'HAS_VULNERABILITY'
-        DISCOVERED = 'DISCOVERED'
-        HAS_ATTRIBUTE = 'HAS_ATTRIBUTE'
-
-    def __init__(self, label: Label, source: 'Node' = None, target: 'Node' = None):
-        self.label = label
-        self.source = source
-        self.target = target
-
-    def to_dict(self):
-        ret = dict(label=self.label.value)
-        if self.source:
-            ret |= dict(source=self.source.to_dict())
-        if self.target:
-            ret |= dict(target=self.target.to_dict())
-        return ret
-
-
-class Node:
-    class Label(Enum): # Chariot.API.My depends upon these labels
-        ASSET = 'Asset'
-        ATTRIBUTE = 'Attribute'
-        RISK = 'Risk'
-        PRESEED = 'Preseed'
-        SEED = 'Seed'
-        TTL = 'TTL'
-
-    def __init__(self, labels: list[Label] = None, filters: list[Filter] = None,
-                 relationships: list[Relationship] = None, params: dict = None):
-        if params:
-            filters = [Filter(params=params)]
-            label = params.get('label', None)
-            if label == None:
-                for filter in filters:
-                    label = get_type_from_key(filter.value)
-            labels = Node.str_to_label(label)
-            relationships = None # Not supported since params does not support relationships
-        self.labels = labels
-        self.filters = filters
-        self.relationships = relationships
-
-    def to_dict(self):
-        ret = dict()
-        if self.labels:
-            ret |= dict(labels=[x.value for x in self.labels])
-        if self.filters:
-            ret |= dict(filters=[x.to_dict() for x in self.filters])
-        if self.relationships:
-            ret |= dict(relationships=[x.to_dict() for x in self.relationships])
-        return ret      
-    
-    @staticmethod
-    def str_to_label(str_of_label: str) -> list[Label]:
-        if not str_of_label:
-            return None
-        for label in Node.Label:
-            if label.name.lower() == str_of_label.lower():
-                return [label]
-        return None
-
-class Query:
-    def __init__(self, node: Node = None, page: int = 0, limit: int = 0, order_by: str = None,
-                 descending: bool = False, params: dict = None):
-        if params:
-            node = Node(params=params)
-            page = params.get('offset', page)
-            limit = params.get('limit', limit)
-            order_by = params.get('orderBy', order_by)
-            descending = params.get('descending', descending)
-        self.node = node
-        self.page = page
-        self.limit = limit
-        self.order_by = order_by
-        self.descending = descending
-
-    def to_dict(self):
-        ret = dict()
-        if self.node:
-            ret |= dict(node=self.node.to_dict())
-        if self.page:
-            ret |= dict(page=self.page)
-        if self.limit:
-            ret |= dict(limit=self.limit)
-        if self.order_by:
-            ret |= dict(orderBy=self.order_by)
-        if self.descending:
-            ret |= dict(descending=self.descending)
-        return ret
-
-
+from praetorian_cli.sdk.model.query import Query
+from praetorian_cli.sdk.model.globals import EXACT_FLAG, DESCENDING_FLAG, GLOBAL_FLAG, Kind
 class Search:
 
     def __init__(self, api):
@@ -210,21 +74,3 @@ def flatten_results(results):
     for key in results.keys():
         flattened.extend(flatten_results(results[key]))
     return flattened
-
-
-# helpers for building graph queries
-ASSET_NODE = [Node.Label.ASSET]
-RISK_NODE = [Node.Label.RISK]
-ATTRIBUTE_NODE = [Node.Label.ATTRIBUTE]
-
-
-def key_equals(key: str):
-    return [Filter(Filter.Field.KEY, Filter.Operator.EQUAL, key)]
-
-
-def risk_of_key(key: str):
-    return Node(RISK_NODE, filters=key_equals(key))
-
-
-def asset_of_key(key: str):
-    return Node(ASSET_NODE, filters=key_equals(key))
